refactor(zone): simplify zone monitor metrics wiring
