--- crates/tempo-zone/src/zone_monitor_metrics.rs
@@ -1,47 +1,0 @@
-//! Metrics for zone block monitoring and batch submission.
-
-use reth_metrics::{
-    Metrics,
-    metrics::{Counter, Gauge, Histogram},
-};
-
-/// Metrics emitted by the zone monitor and batch submitter.
-#[derive(Metrics, Clone)]
-#[metrics(scope = "zone_monitor")]
-pub(crate) struct ZoneMonitorMetrics {
-    /// Most recent zone block observed on L2.
-    pub latest_zone_block_observed: Gauge,
-
-    /// Most recent zone block successfully submitted to L1.
-    pub latest_zone_block_submitted_to_l1: Gauge,
-
-    /// Gap between the latest observed zone block and the latest submitted zone block.
-    pub zone_to_l1_submission_lag_blocks: Gauge,
-
-    /// Number of zone blocks included in a batch submission.
-    pub batch_size_blocks: Histogram,
-
-    /// Number of withdrawals included in a batch submission.
-    pub withdrawals_per_batch: Histogram,
-
-    /// End-to-end latency of a batch submission attempt in seconds.
-    pub batch_submit_latency_seconds: Histogram,
-
-    /// Successful batch submissions.
-    pub batch_submit_success_total: Counter,
-
-    /// Failed batch submissions after exhausting retries.
-    pub batch_submit_failure_total: Counter,
-
-    /// Retry attempts for batch submissions.
-    pub batch_submit_retry_total: Counter,
-
-    /// Successful submissions that used direct anchor mode.
-    pub direct_mode_submissions_total: Counter,
-
-    /// Successful submissions that used ancestry anchor mode.
-    pub ancestry_mode_submissions_total: Counter,
-
-    /// Number of times local monitor state was resynced from the portal.
-    pub resync_from_portal_total: Counter,
-}
